+++ tests/functions/twilio-wrappers/taskrouter.private.test.js
@@ -1,43 +1,34 @@
-import helpers from '../../test-utils/test-helper';
-import axios from 'axios';
+jest.mock('@twilio/flex-plugins-library-utils', () => ({
+  __esModule: true,
+  TaskRouterUtils: jest.fn(),
+}));
+
+import { TaskRouterUtils } from '@twilio/flex-plugins-library-utils';
 
 describe('taskrouter.createTask', () => {
-  const mockCreateTaskFn = jest.fn(() =>
-    Promise.resolve({
-      sid: 'TSxxxxxx',
-      attributes: '{}',
-    }),
-  );
-  const taskRouterTwilioClient = function (updateFunc) {
-    const getWorkspace = (workspaceSid) => ({
-      sid: workspaceSid,
-      tasks: {
-        create: updateFunc,
-      },
-    });
-
-    const mockTaskRouterService = {
-      workspaces: getWorkspace,
-    };
-    return {
-      taskrouter: mockTaskRouterService,
-    };
-  };
-
-  beforeAll(() => {
-    helpers.setup();
-    global.Runtime._addFunction(
-      'twilio-wrappers/retry-handler',
-      './functions/twilio-wrappers/retry-handler.private.js',
-    );
+  afterEach(() => {
+    jest.clearAllMocks();
   });
 
   it('createTask gives success', async () => {
+    TaskRouterUtils.mockImplementation((value) => {
+      return {
+        createTask: jest.fn(() =>
+          Promise.resolve({
+            status: 200,
+            task: {
+              sid: 'TSxxxxxx',
+              attributes: '{}',
+            },
+            success: true,
+          }),
+        ),
+      };
+    });
     const { createTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
 
     const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(mockCreateTaskFn),
+      getTwilioClient: () => () => jest.fn(),
     };
     const payload = {
       context: mockContext,
@@ -57,22 +48,24 @@ describe('taskrouter.createTask', () => {
       task: { sid: 'TSxxxxxx', attributes: {} },
       status: 200,
     });
-    expect(mockCreateTaskFn.mock.calls.length).toBe(1);
-    expect(mockCreateTaskFn.mock.calls[0][0]).toStrictEqual({
-      attributes: '{}',
-      workflowSid: '123',
-      taskChannel: '87678',
-      priority: 1,
-      timeout: 898,
-    });
   });
 
-  it('createTask gives error due to invalid taskChannel', async () => {
+  it('createTask gives error', async () => {
+    TaskRouterUtils.mockImplementation((value) => {
+      return {
+        createTask: jest.fn(() =>
+          Promise.reject({
+            success: false,
+            status: 400,
+            message: 'Mock Error Message',
+          }),
+        ),
+      };
+    });
     const { createTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
 
     const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(updateFunc),
+      getTwilioClient: () => () => jest.fn(),
     };
     const payload = {
       context: mockContext,
@@ -85,72 +78,41 @@ describe('taskrouter.createTask', () => {
       attributes: {},
       attempts: 0,
     };
-    await createTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain taskChannel string');
-    });
-  });
-  it('createTask gives error due to invalid attributes', async () => {
-    const { createTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
 
-    const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(updateFunc),
-    };
-    const payload = {
-      context: mockContext,
-      workflowSid: '123',
-      taskChannel: '87678',
-      priority: {
-        overriddenPriority: 1,
-      },
-      timeOut: { overriddenTimeout: 898 },
-      attributes: '',
-      attempts: 0,
-    };
-    await createTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain attributes object');
+    const errTask = await createTask({ ...payload });
+
+    expect(errTask).toEqual({
+      success: false,
+      status: 400,
+      message: 'Mock Error Message',
     });
   });
 });
 
 describe('taskrouter.fetchTask', () => {
-  const mockFetchTaskFn = jest.fn(() =>
-    Promise.resolve({
-      sid: 'TSxxxxxx',
-      attributes: '{}',
-    }),
-  );
-  const taskRouterTwilioClient = function (fetchFunc) {
-    const getWorkspace = (workspaceSid) => ({
-      sid: workspaceSid,
-      tasks: (_taskSid) => ({
-        fetch: fetchFunc,
-      }),
-    });
-
-    const mockTaskRouterService = {
-      workspaces: getWorkspace,
-    };
-    return {
-      taskrouter: mockTaskRouterService,
-    };
-  };
-
-  beforeAll(() => {
-    helpers.setup();
-
-    global.Runtime._addFunction(
-      'twilio-wrappers/retry-handler',
-      './functions/twilio-wrappers/retry-handler.private.js',
-    );
+  afterEach(() => {
+    jest.clearAllMocks();
   });
 
   it('fetchTask gives success', async () => {
+    TaskRouterUtils.mockImplementation((value) => {
+      return {
+        fetchTask: jest.fn(() =>
+          Promise.resolve({
+            status: 200,
+            task: {
+              sid: 'TSxxxxxx',
+              attributes: '{}',
+            },
+            success: true,
+          }),
+        ),
+      };
+    });
     const { fetchTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
 
     const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(mockFetchTaskFn),
+      getTwilioClient: () => () => jest.fn(),
     };
     const payload = {
       context: mockContext,
@@ -165,89 +127,68 @@ describe('taskrouter.fetchTask', () => {
       task: { sid: 'TSxxxxxx', attributes: {} },
       status: 200,
     });
-    expect(mockFetchTaskFn.mock.calls.length).toBe(1);
   });
 
-  it('fetchTask gives error due to invalid attempts', async () => {
+  it('fetchTask gives error', async () => {
+    TaskRouterUtils.mockImplementation((value) => {
+      return {
+        fetchTask: jest.fn(() =>
+          Promise.reject({
+            success: false,
+            status: 400,
+            message: 'Mock Error Message',
+          }),
+        ),
+      };
+    });
     const { fetchTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
 
     const mockContext = {
-      context: mockContext,
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(updateFunc),
+      getTwilioClient: () => () => jest.fn(),
     };
     const payload = {
+      context: mockContext,
       taskSid: 'TSxxxxxx',
       attempts: '0',
     };
-    await fetchTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain the number of attempts');
-    });
-  });
 
-  it('fetchTask gives error due to invalid taskSid', async () => {
-    const { fetchTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
+    const errTask = await fetchTask({ ...payload });
 
-    const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(updateFunc),
-    };
-    const payload = {
-      context: mockContext,
-      taskSid: 123,
-      attempts: 0,
-    };
-    await fetchTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain the taskSid string');
-    });
-  });
-
-  it('fetchTask gives error due to invalid context', async () => {
-    const { fetchTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
-
-    const payload = {
-      context: 'mockContext',
-      taskSid: '123',
-      attempts: 0,
-    };
-    await fetchTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain reason context object');
+    expect(errTask).toEqual({
+      success: false,
+      status: 400,
+      message: 'Mock Error Message',
     });
   });
 });
 
 describe('taskrouter.updateTaskAttributes', () => {
-  jest.mock('axios', () => ({
-    post: jest.fn((_url, _body) => {
-      return new Promise((resolve) => {
-        resolve({
-          data: {
-            attributes: '{"attr1": "mockValue","attr2": "mockValue"}',
-          },
-        });
-      });
-    }),
-    get: jest.fn((_url, _body) => {
-      return new Promise((resolve) => {
-        resolve({ data: { attributes: '{ "attr1": "mockValue" }' }, headers: { etag: '{"attr1":"value1"}' } });
-      });
-    }),
-  }));
-  beforeAll(() => {
-    helpers.setup();
-    global.Runtime._addFunction(
-      'twilio-wrappers/retry-handler',
-      './functions/twilio-wrappers/retry-handler.private.js',
-    );
+  afterEach(() => {
+    jest.clearAllMocks();
   });
-
   it('updateTaskAttributes gives success', async () => {
+    TaskRouterUtils.mockImplementation((value) => {
+      return {
+        updateTaskAttributes: jest.fn(() =>
+          Promise.resolve({
+            status: 200,
+            task: {
+              attributes: '{ "attr1": "mockValue", "attr2": "mockValue" }',
+            },
+            success: true,
+          }),
+        ),
+      };
+    });
     const { updateTaskAttributes } = require('../../../functions/twilio-wrappers/taskrouter.private');
+    const mockContext = {
+      getTwilioClient: () => () => jest.fn(),
+    };
     const payload = {
-      attributesUpdate: '{"attr2": "mockValue"}',
+      context: mockContext,
+      attributesUpdate: { attr1: 'mockValue', attr2: 'mockValue' },
       taskSid: 'TSxxxxxx',
       attempts: 0,
-      context: {}
     };
 
     const task = await updateTaskAttributes({ ...payload });
@@ -259,81 +200,62 @@ describe('taskrouter.updateTaskAttributes', () => {
     });
   });
 
-  it('updateTaskAttributes gives error due to invalid attempts', async () => {
+  it('updateTaskAttributes gives error', async () => {
+    TaskRouterUtils.mockImplementation((value) => {
+      return {
+        updateTaskAttributes: jest.fn(() =>
+          Promise.reject({
+            success: false,
+            status: 400,
+            message: 'Mock Error Message',
+          }),
+        ),
+      };
+    });
     const { updateTaskAttributes } = require('../../../functions/twilio-wrappers/taskrouter.private');
+    const mockContext = {
+      getTwilioClient: () => () => jest.fn(),
+    };
     const payload = {
-      attributesUpdate: '{"attr2": "mockValue"}',
+      context: mockContext,
+      attributesUpdate: '{}',
       taskSid: 'TSxxxxxx',
       attempts: '0',
     };
-    await updateTaskAttributes({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain the number of attempts');
-    });
-  });
 
-  it('updateTaskAttributes gives error due to invalid taskSid', async () => {
-    const { updateTaskAttributes } = require('../../../functions/twilio-wrappers/taskrouter.private');
-    const payload = {
-      attributesUpdate: '{"attr2": "mockValue"}',
-      taskSid: 123,
-      attempts: 0,
-    };
-    await updateTaskAttributes({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain the taskSid string');
-    });
-  });
+    const errTask = await updateTaskAttributes({ ...payload });
 
-  it('updateTaskAttributes gives error due to invalid attributesUpdate', async () => {
-    const { updateTaskAttributes } = require('../../../functions/twilio-wrappers/taskrouter.private');
-
-    const payload = {
-      attributesUpdate: { attr2: 'mockValue' },
-      taskSid: '123',
-      attempts: 0,
-    };
-    await updateTaskAttributes({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain attributesUpdate JSON string');
+    expect(errTask).toEqual({
+      success: false,
+      status: 400,
+      message: 'Mock Error Message',
     });
   });
 });
 
 describe('taskrouter.updateTask', () => {
-  const mockUpdateTaskFn = jest.fn(() =>
-    Promise.resolve({
-      sid: 'TSxxxxxx',
-      attributes: '{}',
-    }),
-  );
-  const taskRouterTwilioClient = function (updateFunc) {
-    const getWorkspace = (workspaceSid) => ({
-      sid: workspaceSid,
-      tasks: (_taskSid) => ({
-        update: updateFunc,
-      }),
-    });
-
-    const mockTaskRouterService = {
-      workspaces: getWorkspace,
-    };
-    return {
-      taskrouter: mockTaskRouterService,
-    };
-  };
-
-  beforeAll(() => {
-    helpers.setup();
-    global.Runtime._addFunction(
-      'twilio-wrappers/retry-handler',
-      './functions/twilio-wrappers/retry-handler.private.js',
-    );
+  afterEach(() => {
+    jest.clearAllMocks();
   });
-
   it('updateTask gives success', async () => {
+    TaskRouterUtils.mockImplementation((value) => {
+      return {
+        updateTask: jest.fn(() =>
+          Promise.resolve({
+            status: 200,
+            task: {
+              sid: 'TSxxxxxx',
+              attributes: '{}',
+            },
+            success: true,
+          }),
+        ),
+      };
+    });
     const { updateTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
 
     const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(mockUpdateTaskFn),
+      getTwilioClient: () => () => jest.fn(),
     };
     const payload = {
       context: mockContext,
@@ -349,16 +271,24 @@ describe('taskrouter.updateTask', () => {
       status: 200,
       task: { sid: 'TSxxxxxx', attributes: {} },
     });
-    expect(mockUpdateTaskFn.mock.calls.length).toBe(1);
-    expect(mockUpdateTaskFn.mock.calls[0][0]).toStrictEqual({});
   });
 
-  it('updateTask gives error due to invalid attempts', async () => {
+  it('updateTask gives error', async () => {
+    TaskRouterUtils.mockImplementation((value) => {
+      return {
+        updateTask: jest.fn(() =>
+          Promise.reject({
+            success: false,
+            status: 400,
+            message: 'Mock Error Message',
+          }),
+        ),
+      };
+    });
     const { updateTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
 
     const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(updateFunc),
+      getTwilioClient: () => () => jest.fn(),
     };
     const payload = {
       context: mockContext,
@@ -366,55 +296,11 @@ describe('taskrouter.updateTask', () => {
       updateParams: {},
       attempts: '0',
     };
-    await updateTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain the number of attempts');
-    });
-  });
-  it('updateTask gives error due to invalid taskSid', async () => {
-    const { updateTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
-
-    const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(updateFunc),
-    };
-    const payload = {
-      context: mockContext,
-      taskSid: 12345,
-      updateParams: {},
-      attempts: 0,
-    };
-    await updateTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain the taskSid string');
-    });
-  });
-  it('updateTask gives error due to invalid updateParams', async () => {
-    const { updateTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
-
-    const mockContext = {
-      PATH: 'mockPath',
-      getTwilioClient: () => taskRouterTwilioClient(updateFunc),
-    };
-    const payload = {
-      context: mockContext,
-      taskSid: 'TSxxxxxx',
-      updateParams: '{}',
-      attempts: 0,
-    };
-    await updateTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain updateParams object');
-    });
-  });
-  it('updateTask gives error due to invalid context', async () => {
-    const { updateTask } = require('../../../functions/twilio-wrappers/taskrouter.private');
-
-    const payload = {
-      context: '',
-      taskSid: 'TSxxxxxx',
-      updateParams: {},
-      attempts: 0,
-    };
-    await updateTask({ ...payload }).catch((err) => {
-      expect(err).toMatch('Invalid parameters object passed. Parameters must contain reason context object');
+    const errTask = await updateTask({ ...payload });
+    expect(errTask).toEqual({
+      success: false,
+      status: 400,
+      message: 'Mock Error Message',
     });
   });
 });