+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## version 1.0.1
+
+- Introduced the usage of flex-ui-telemetry package instead of individual classes.
+- Introduced the usage of flex-plugins-library-utils package, which is a helper package with all the common twilio-functions readily used amongst the plugins.
+
 ## version 1.0.0
 
 This feature adds a new "Call Agent" section to the outbound dialpad allowing an agent to directly call another agent. In this section, there is an autocomplete dropdown where you can select the agent you want to call.